+++ src/test/java/lotto/util/RandomNumGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 import static camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest;
 import static org.assertj.core.api.Assertions.assertThat;
